ADDED: static indexOfProperty() to find String.indexOf() of property names in Strings, which are always followed by a ":".
Removes need to add ":" throughout the code

incrementHashMap() counts instances of "key" and increments the count kept in HashMap.value . If "key" is not yet a key it is created with a starting value of 1

CHANGED:
Construction of boundedOutput in boundedStringByValue() altered with '+ ":" ' to accommodate use of Reporter.indexOfProperty 
Replaced String.indexOf() with indexOfProperty()

DEBUGGED:
CompareValue() called in boundedStringByValue() to check value of propertyName was using startIndex as a starting index, but this was index of bound (usually "contact") in string. 
Overloaded compareValue() to not need starting index.