ADDED: initInfectedStatus() Handles initialising of infectedStatus, ensures consistency with the infectedStatus of the Sites.
Extra code in setStatusHIV() to ensure that prepStatus and antiViralStatus are consistent.

CHANGED:
static birthMSM() now returns MSM instead of Object.
InitStatus now takes int argument startAge so newly born MSM have statusHIV false and infectedStatus false. 
Extra code in setStatusHIV() to ensure that prepStatus and antiViralStatus are consistent. 
Extra line in setAntiViralStatus() to ensure it is not true unless statusHIV is already true, for consistency. 
Extra line in setPrepStatus() to ensure it is not true unless statusHIV is already false, for consistency.